fixing more memory leaks and some errors in the wrapper detected from new integration tests
