+++ .gitignore
@@ -1,3 +1,51 @@
 /build/
 /tls-proxy/.git/
 /feartls/.git/
+/tls-proxy/License.md
+/tls-proxy/.gitignore
+
+# Python
+__pycache__/
+*.py[cod]
+*$py.class
+*.so
+.Python
+
+# Virtual environments
+venv/
+env/
+ENV/
+.venv
+
+# Python package distribution
+*.egg-info/
+dist/
+*.egg
+
+# PyInstaller
+*.manifest
+*.spec
+
+# Unit test / coverage
+.pytest_cache/
+.coverage
+htmlcov/
+*.cover
+
+# IDEs
+.vscode/
+.idea/
+*.swp
+*.swo
+*~
+
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# Python test scripts (from test suite)
+test_script_*.py
+test_AsyncDataProcessing_*.py
+test_single_interp_script*.py
+
+# Test config directories
+test_config_*/
